Convert the project to a Ledger standard app
