+++ src/apdu/dispatcher.c
@@ -28,9 +28,9 @@
 #include "../constants.h"
 #include "../globals.h"
 #include "../types.h"
-#include "../io.h"
+#include "io.h"
 #include "../sw.h"
-#include "../common/buffer.h"
+#include "buffer.h"
 #include "../handler/get_version.h"
 #include "../handler/get_app_name.h"
 #include "../handler/get_public_key.h"
