--- src/common/buffer.c
@@ -1,158 +1,0 @@
-/*****************************************************************************
- * MIT License
- *
- * Copyright (c) 2023 coderofstuff
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- *****************************************************************************/
-
-#include <stdint.h>   // uint*_t
-#include <stddef.h>   // size_t
-#include <stdbool.h>  // bool
-#include <string.h>   // memmove
-
-#include "buffer.h"
-#include "read.h"
-#include "bip32.h"
-
-bool buffer_can_read(const buffer_t *buffer, size_t n) {
-    return buffer->size - buffer->offset >= n;
-}
-
-bool buffer_seek_set(buffer_t *buffer, size_t offset) {
-    if (offset > buffer->size) {
-        return false;
-    }
-
-    buffer->offset = offset;
-
-    return true;
-}
-
-bool buffer_seek_cur(buffer_t *buffer, size_t offset) {
-    if (buffer->offset + offset < buffer->offset ||  // overflow
-        buffer->offset + offset > buffer->size) {    // exceed buffer size
-        return false;
-    }
-
-    buffer->offset += offset;
-
-    return true;
-}
-
-bool buffer_seek_end(buffer_t *buffer, size_t offset) {
-    if (offset > buffer->size) {
-        return false;
-    }
-
-    buffer->offset = buffer->size - offset;
-
-    return true;
-}
-
-bool buffer_read_u8(buffer_t *buffer, uint8_t *value) {
-    if (!buffer_can_read(buffer, 1)) {
-        *value = 0;
-
-        return false;
-    }
-
-    *value = buffer->ptr[buffer->offset];
-    buffer_seek_cur(buffer, 1);
-
-    return true;
-}
-
-bool buffer_read_u16(buffer_t *buffer, uint16_t *value, endianness_t endianness) {
-    if (!buffer_can_read(buffer, 2)) {
-        *value = 0;
-
-        return false;
-    }
-
-    *value = ((endianness == BE) ? read_u16_be(buffer->ptr, buffer->offset)
-                                 : read_u16_le(buffer->ptr, buffer->offset));
-
-    buffer_seek_cur(buffer, 2);
-
-    return true;
-}
-
-bool buffer_read_u32(buffer_t *buffer, uint32_t *value, endianness_t endianness) {
-    if (!buffer_can_read(buffer, 4)) {
-        *value = 0;
-
-        return false;
-    }
-
-    *value = ((endianness == BE) ? read_u32_be(buffer->ptr, buffer->offset)
-                                 : read_u32_le(buffer->ptr, buffer->offset));
-
-    buffer_seek_cur(buffer, 4);
-
-    return true;
-}
-
-bool buffer_read_u64(buffer_t *buffer, uint64_t *value, endianness_t endianness) {
-    if (!buffer_can_read(buffer, 8)) {
-        *value = 0;
-
-        return false;
-    }
-
-    *value = ((endianness == BE) ? read_u64_be(buffer->ptr, buffer->offset)
-                                 : read_u64_le(buffer->ptr, buffer->offset));
-
-    buffer_seek_cur(buffer, 8);
-
-    return true;
-}
-
-bool buffer_read_bip32_path(buffer_t *buffer, uint32_t *out, size_t out_len) {
-    if (!bip32_path_read(buffer->ptr + buffer->offset,
-                         buffer->size - buffer->offset,
-                         out,
-                         out_len)) {
-        return false;
-    }
-
-    buffer_seek_cur(buffer, sizeof(*out) * out_len);
-
-    return true;
-}
-
-bool buffer_copy(const buffer_t *buffer, uint8_t *out, size_t out_len) {
-    if (buffer->size - buffer->offset > out_len) {
-        return false;
-    }
-
-    memmove(out, buffer->ptr + buffer->offset, buffer->size - buffer->offset);
-
-    return true;
-}
-
-bool buffer_move(buffer_t *buffer, uint8_t *out, size_t out_len) {
-    if (!buffer_copy(buffer, out, out_len)) {
-        return false;
-    }
-
-    buffer_seek_cur(buffer, out_len);
-
-    return true;
-}
