+++ src/handler/get_public_key.c
@@ -27,15 +27,15 @@
 #include <string.h>   // memset, explicit_bzero
 
 #include "os.h"
-#include "cx.h"
+#include "crypto_helpers.h"
 
 #include "get_public_key.h"
 #include "../globals.h"
 #include "../types.h"
-#include "../io.h"
+#include "io.h"
 #include "../sw.h"
 #include "../crypto.h"
-#include "../common/buffer.h"
+#include "buffer.h"
 #include "../ui/display.h"
 #include "../helper/send_response.h"
 
@@ -44,8 +44,7 @@ int handler_get_public_key(buffer_t *cdata, bool display) {
     G_context.req_type = CONFIRM_ADDRESS;
     G_context.state = STATE_NONE;
 
-    cx_ecfp_private_key_t private_key = {0};
-    cx_ecfp_public_key_t public_key = {0};
+    uint8_t raw_pubkey[65] = {0};
 
     if (!buffer_read_u8(cdata, &G_context.bip32_path_len) ||
         !buffer_read_bip32_path(cdata, G_context.bip32_path, (size_t) G_context.bip32_path_len)) {
@@ -69,18 +68,14 @@ int handler_get_public_key(buffer_t *cdata, bool display) {
         return io_send_sw(SW_WRONG_BIP32_TYPE);
     }
 
-    // derive private key according to BIP32 path
-    int error = crypto_derive_private_key(&private_key,
-                                          G_context.pk_info.chain_code,
-                                          G_context.bip32_path,
-                                          G_context.bip32_path_len);
-    if (error != 0) {
-        return io_send_sw(error);
-    }
-    // generate corresponding public key
-    crypto_init_public_key(&private_key, &public_key, G_context.pk_info.raw_public_key);
-    // reset private key
-    explicit_bzero(&private_key, sizeof(private_key));
+    bip32_derive_get_pubkey_256(CX_CURVE_256K1,
+                                G_context.bip32_path,
+                                G_context.bip32_path_len,
+                                raw_pubkey,
+                                G_context.pk_info.chain_code,
+                                NULL);
+
+    memmove(G_context.pk_info.raw_public_key, raw_pubkey + 1, 64);
 
     if (display) {
         return ui_display_address();
