+++ src/handler/get_version.c
@@ -28,10 +28,10 @@
 #include "get_version.h"
 #include "../globals.h"
 #include "../constants.h"
-#include "../io.h"
+#include "io.h"
 #include "../sw.h"
 #include "../types.h"
-#include "common/buffer.h"
+#include "buffer.h"
 
 int handler_get_version() {
     _Static_assert(APPVERSION_LEN == 3, "Length of (MAJOR || MINOR || PATCH) must be 3!");
@@ -42,11 +42,8 @@ int handler_get_version() {
     _Static_assert(PATCH_VERSION >= 0 && PATCH_VERSION <= UINT8_MAX,
                    "PATCH version must be between 0 and 255!");
 
-    return io_send_response(
-        &(const buffer_t){.ptr = (uint8_t[APPVERSION_LEN]){(uint8_t) MAJOR_VERSION,
-                                                           (uint8_t) MINOR_VERSION,
-                                                           (uint8_t) PATCH_VERSION},
-                          .size = APPVERSION_LEN,
-                          .offset = 0},
+    return io_send_response_pointer(
+        (uint8_t[APPVERSION_LEN]){(uint8_t) MAJOR_VERSION, (uint8_t) MINOR_VERSION, (uint8_t) PATCH_VERSION},
+        APPVERSION_LEN,
         SW_OK);
 }
