--- src/io.h
@@ -1,88 +1,0 @@
-/*****************************************************************************
- * MIT License
- *
- * Copyright (c) 2023 coderofstuff
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- *****************************************************************************/
-#pragma once
-
-#include <stdint.h>
-
-#include "ux.h"
-#include "os_io_seproxyhal.h"
-
-#include "types.h"
-#include "common/buffer.h"
-
-#ifdef HAVE_BAGL
-void io_seproxyhal_display(const bagl_element_t *element);
-#endif  // HAVE_BAGL
-
-/**
- * IO callback called when an interrupt based channel has received
- * data to be processed.
- *
- * @return 1 if success, 0 otherwise.
- *
- */
-uint8_t io_event(uint8_t channel);
-
-uint16_t io_exchange_al(uint8_t channel, uint16_t tx_len);
-
-/**
- * Initialize the APDU I/O state.
- *
- * This function must be called before calling any other I/O function.
- */
-void io_init(void);
-
-/**
- * Receive APDU command in G_io_apdu_buffer.
- *
- * @return zero or positive integer if success, -1 otherwise.
- *
- */
-int io_recv_command(void);
-
-/**
- * Send APDU response (response data + status word) by filling
- * G_io_apdu_buffer.
- *
- * @param[in] rdata
- *   Buffer with APDU response data.
- * @param[in] sw
- *   Status word of APDU response.
- *
- * @return zero or positive integer if success, -1 otherwise.
- *
- */
-int io_send_response(const buffer_t *rdata, uint16_t sw);
-
-/**
- * Send APDU response (only status word) by filling
- * G_io_apdu_buffer.
- *
- * @param[in] sw
- *   Status word of APDU response.
- *
- * @return zero or positive integer if success, -1 otherwise.
- *
- */
-int io_send_sw(uint16_t sw);
