+++ src/types.h
@@ -31,15 +31,6 @@
 #include "common/bip32.h"
 
 /**
- * Enumeration for the status of IO.
- */
-typedef enum {
-    READY,     /// ready for new event
-    RECEIVED,  /// data received
-    WAITING    /// waiting
-} io_state_e;
-
-/**
  * Enumeration with expected INS of APDU commands.
  */
 typedef enum {
