+++ src/ui/nbgl_display_address.c
@@ -36,7 +36,7 @@
 #include "display.h"
 #include "constants.h"
 #include "../globals.h"
-#include "../io.h"
+#include "io.h"
 #include "../sw.h"
 #include "../address.h"
 #include "action/validate.h"
